+++ test/fixtures/duplicateIds.json
@@ -16,6 +16,11 @@
   },
   "fieldMapping": [
     {
+      "type": "id",
+      "name": "wikidata",
+      "desc": "Wikidata identifier"
+    },
+    {
       "type": "property",
       "name": "label_en",
       "desc": "Region name (English)"
@@ -45,6 +50,11 @@
     "zh": "艾辛格"
   },
   "fieldMapping": [
+    {
+      "type": "id",
+      "name": "wikidata",
+      "desc": "Wikidata identifier"
+    },
     {
       "type": "property",
       "name": "label_en",
