+++ test/fixtures/fieldInfo.json
@@ -1,0 +1,18 @@
+{
+  "wikidata": {
+    "wikidata": "Q43649390",
+    "i18n": {
+      "de": "Wikidata-Kennung",
+      "en": "Wikidata identifier",
+      "zh": "维基数据标识符"
+    }
+  },
+  "name": {
+    "wikidata": "P2561",
+    "i18n": {
+      "de": "name",
+      "en": "name",
+      "zh": "名称"
+    }
+  }
+}
